ml/makeup_advisor/requirements.txt
tensorflow==2.14.0
opencv-python
mtcnn
numpy
pillow
scikit-learn
pandas
matplotlib
tqdm
